Fix deepEquals when one param is object and other is null
When a variable is equal to null, typeof returns "object". When one
of the values is an object and the other is null, this scenario
would throw an error attempting to execute Object.keys(null).

This commit fixes the issue by checking for nulls if both values
are of type "object".

This commit also fixes linting issues in deepEquals.spec.js